pdf_labeler/requirements.txt
Flask==2.3.3
pdfplumber==0.10.3
PyPDF2==3.0.1
pandas==2.1.4
openai==1.12.0
python-dateutil==2.8.2

